asceticCCFAssessment

Perform the assessment via cross-validation of the model inferred by the ASCETIC framework on single samples (using CCF) datasets.

asceticCCFAssessment(
  inference,
  niterations = 100,
  vafDataset = NULL,
  nsampling = 100
)

Arguments

inference

Model inferred with ASCETIC using either the function asceticCCFResampling or the function asceticCCF

niterations

Number of cross-validation iterations to be performed for a robust assessment of ASCETIC model. Higher values lead to improved estimates, but require higher computational burden; default value is 100.

vafDataset

R data.frame with 8 columns: 1) SAMPLE_ID, sample name. 2) GENE_ID, gene name. 3) REF_COUNT, total counts for reference allele. 4) ALT_COUNT, total counts for alternate allele. 5) COPY_NUMBER, total copy number estimate. 6) NORMAL_PLOIDY, ploidy for normal sample; this is either 1 for mutations on sex chromosomes or 2. 7) VAF_ESTIMATE, variant allele frequency (VAF) estimate. 8) CCF_ESTIMATE, cancer cell fraction (CCF) estimate. Values reported in vafDataset must be consistent with the ones reported in dataset and ccfDataset. If vafDataset is not provided, sampling cannot be performed.

nsampling

Number of re-sampling to be performed for a robust estimation of the agony ranking. Higher values lead to improved estimates, but require higher computational burden; default value is 100.

Value

A list of 3 elements for which the estimate by cross-validation is performed: 1) rankingEstimate, ranking among mutations estimated by agony. Lower rankings correspond to early mutations. This is returned only if nsampling > 0. 2) poset, partially order set among mutations estimated by ASCETIC from the agony ranking. 3) inference, inferred ASCETIC evolutionary model for each selected regularization.

Examples

set.seed(12345)
data(datasetExampleSingleSamples)
data(ccfDatasetExampleSingleSamples)
data(vafDatasetExampleSingleSamples)
resExampleSingleSamplesResampling <- asceticCCFResampling(
                                               dataset = datasetExampleSingleSamples,
                                               ccfDataset = ccfDatasetExampleSingleSamples,
                                               vafDataset = vafDatasetExampleSingleSamples,
                                               nsampling = 3,
                                               regularization = c("aic","bic"),
                                               command = "hc",
                                               restarts = 0 )
#> 0 
#> 0.3333333 
#> 0.6666667 
#> 1 
resExampleCCFAssessment <- asceticCCFAssessment(
                                       inference = resExampleSingleSamplesResampling,
                                       niterations = 3,
                                       vafDataset = vafDatasetExampleSingleSamples,
                                       nsampling = 3)
#> Starting cross-validation... 
#> 0 
#> 0.3333333 
#> 0.6666667 
#> 1 
#> Cross-validation progress:  0.3333333 
#> 0 
#> 0.3333333 
#> 0.6666667 
#> 1 
#> Warning: variable 2 in the data has levels that are not observed in the data.
#> Warning: variable 2 in the data has levels that are not observed in the data.
#> Cross-validation progress:  0.6666667 
#> 0 
#> 0.3333333 
#> 0.6666667 
#> 1 
#> Cross-validation progress:  1